Swap Esc and Del on Thumb Clusters
After much usage, I've decided that swapping the Esc and Del I had on
the thumb clusters is for the best; on most keyboards, Esc is in the top
left and Del is either Fn + Bkspc or is on the top right (in the nav
cluster to the left of the numpad). Having these reversed was originally
for ease of perma-deleting (with Shift + Delete), but I have decided
that it is no longer worth it.